local_tools/ants_tool_maker_tutorial/pyproject.toml
[project]
name = "ants-tool-maker-tutorial"
version = "1.0"
description = ""
authors = [
    {name = "Marc Ramos",email = "marc@qmenta.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
antspyx = "0.6.1"
qmenta-sdk-lib = "^2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
